Port Journeymap and racks etc to 1.17 (#7877)
Port Add "universitylibrary" placement-helper blueprints for medieval styles #7844
Port Fixes rendering of rack window #7870
Port Journeymap integration #7858
Minimap doesn't show colony border for some reason, but it seems like a JM bug since it works in 1.16
Clean up dependency version properties and automate mods.toml
Bump min forge version both for JM compatibility and to ensure log4j-safe version

+++ src/api/java/com/minecolonies/api/util/Utils.java
@@ -166,7 +166,7 @@ public final class Utils
      * @param flag Flag to check whether it is set or not.
      * @return True if flag is set, otherwise false.
      */
-    public static boolean testFlag(final int data, final int flag)
+    public static boolean testFlag(final long data, final long flag)
     {
         return mask(data, flag) == flag;
     }
@@ -178,7 +178,7 @@ public final class Utils
      * @param mask Mask to check.
      * @return Byte in which both data bits and mask bits are set.
      */
-    public static int mask(final int data, final int mask)
+    public static long mask(final long data, final long mask)
     {
         return data & mask;
     }
@@ -190,7 +190,7 @@ public final class Utils
      * @param flag Flag to set.
      * @return Data with flags set.
      */
-    public static int setFlag(final int data, final int flag)
+    public static long setFlag(final long data, final long flag)
     {
         return data | flag;
     }
@@ -202,7 +202,7 @@ public final class Utils
      * @param flag Flag to remove.
      * @return Data with flag unset.
      */
-    public static int unsetFlag(final int data, final int flag)
+    public static long unsetFlag(final long data, final long flag)
     {
         return data & ~flag;
     }
@@ -214,7 +214,7 @@ public final class Utils
      * @param flag Flag to toggle.
      * @return Data with flag toggled.
      */
-    public static int toggleFlag(final int data, final int flag)
+    public static long toggleFlag(final long data, final long flag)
     {
         return data ^ flag;
     }
